% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{accept_certificate_transfer}
\alias{accept_certificate_transfer}
\title{Accepts a pending certificate transfer}
\usage{
accept_certificate_transfer(certificateId, setAsActive = NULL)
}
\arguments{
\item{certificateId}{[required] The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)}

\item{setAsActive}{Specifies whether the certificate is active.}
}
\description{
Accepts a pending certificate transfer. The default state of the certificate is INACTIVE.
}
\details{
To check for pending certificate transfers, call ListCertificates to enumerate your certificates.
}
\section{Accepted Parameters}{
\preformatted{accept_certificate_transfer(
  certificateId = "string",
  setAsActive = TRUE|FALSE
)
}
}

