% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{add_thing_to_billing_group}
\alias{add_thing_to_billing_group}
\title{Adds a thing to a billing group}
\usage{
add_thing_to_billing_group(billingGroupName = NULL,
  billingGroupArn = NULL, thingName = NULL, thingArn = NULL)
}
\arguments{
\item{billingGroupName}{The name of the billing group.}

\item{billingGroupArn}{The ARN of the billing group.}

\item{thingName}{The name of the thing to be added to the billing group.}

\item{thingArn}{The ARN of the thing to be added to the billing group.}
}
\description{
Adds a thing to a billing group.
}
\section{Accepted Parameters}{
\preformatted{add_thing_to_billing_group(
  billingGroupName = "string",
  billingGroupArn = "string",
  thingName = "string",
  thingArn = "string"
)
}
}

