% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{add_thing_to_thing_group}
\alias{add_thing_to_thing_group}
\title{Adds a thing to a thing group}
\usage{
add_thing_to_thing_group(thingGroupName = NULL, thingGroupArn = NULL,
  thingName = NULL, thingArn = NULL, overrideDynamicGroups = NULL)
}
\arguments{
\item{thingGroupName}{The name of the group to which you are adding a thing.}

\item{thingGroupArn}{The ARN of the group to which you are adding a thing.}

\item{thingName}{The name of the thing to add to a group.}

\item{thingArn}{The ARN of the thing to add to a group.}

\item{overrideDynamicGroups}{Override dynamic thing groups with static thing groups when 10-group limit is reached. If a thing belongs to 10 thing groups, and one or more of those groups are dynamic thing groups, adding a thing to a static group removes the thing from the last dynamic group.}
}
\description{
Adds a thing to a thing group.
}
\section{Accepted Parameters}{
\preformatted{add_thing_to_thing_group(
  thingGroupName = "string",
  thingGroupArn = "string",
  thingName = "string",
  thingArn = "string",
  overrideDynamicGroups = TRUE|FALSE
)
}
}

