% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{attach_policy}
\alias{attach_policy}
\title{Attaches a policy to the specified target}
\usage{
attach_policy(policyName, target)
}
\arguments{
\item{policyName}{The name of the policy to attach.}

\item{target}{The identity to which the policy is attached.}
}
\description{
Attaches a policy to the specified target.
}
\examples{

}
