% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{attach_security_profile}
\alias{attach_security_profile}
\title{Associates a Device Defender security profile with a thing group or with this account}
\usage{
attach_security_profile(securityProfileName, securityProfileTargetArn)
}
\arguments{
\item{securityProfileName}{[required] The security profile that is attached.}

\item{securityProfileTargetArn}{[required] The ARN of the target (thing group) to which the security profile is attached.}
}
\description{
Associates a Device Defender security profile with a thing group or with this account. Each thing group or account can have up to five security profiles associated with it.
}
\section{Accepted Parameters}{
\preformatted{attach_security_profile(
  securityProfileName = "string",
  securityProfileTargetArn = "string"
)
}
}

