% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{cancel_certificate_transfer}
\alias{cancel_certificate_transfer}
\title{Cancels a pending transfer for the specified certificate}
\usage{
cancel_certificate_transfer(
  certificateId = "string"
)
}
\arguments{
\item{certificateId}{[required] The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)}
}
\description{
Cancels a pending transfer for the specified certificate.
}
\details{
\strong{Note} Only the transfer source account can use this operation to cancel a transfer. (Transfer destinations can use RejectCertificateTransfer instead.) After transfer, AWS IoT returns the certificate to the source account in the INACTIVE state. After the destination account has accepted the transfer, the transfer cannot be cancelled.

After a certificate transfer is cancelled, the status of the certificate changes from PENDING\_TRANSFER to INACTIVE.
}
\examples{

}
