% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{cancel_job}
\alias{cancel_job}
\title{Cancels a job}
\usage{
cancel_job(jobId, comment = NULL, force = NULL)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}

\item{comment}{An optional comment string describing why the job was canceled.}

\item{force}{(Optional) If \code{true} job executions with status "IN\_PROGRESS" and "QUEUED" are canceled, otherwise only job executions with status "QUEUED" are canceled. The default is \code{false}.

Canceling a job which is "IN\_PROGRESS", will cause a device which is executing the job to be unable to update the job execution status. Use caution and ensure that each device executing a job which is canceled is able to recover to a valid state.}
}
\description{
Cancels a job.
}
\section{Accepted Parameters}{
\preformatted{cancel_job(
  jobId = "string",
  comment = "string",
  force = TRUE|FALSE
)
}
}

