% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{clear_default_authorizer}
\alias{clear_default_authorizer}
\title{Clears the default authorizer}
\usage{
clear_default_authorizer()
}
\description{
Clears the default authorizer.
}
\section{Accepted Parameters}{
\preformatted{clear_default_authorizer()
}
}

