% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_authorizer}
\alias{create_authorizer}
\title{Creates an authorizer}
\usage{
create_authorizer(authorizerName, authorizerFunctionArn, tokenKeyName,
  tokenSigningPublicKeys, status = NULL)
}
\arguments{
\item{authorizerName}{The authorizer name.}

\item{authorizerFunctionArn}{The ARN of the authorizer's Lambda function.}

\item{tokenKeyName}{The name of the token key used to extract the token from the HTTP headers.}

\item{tokenSigningPublicKeys}{The public keys used to verify the digital signature returned by your custom authentication service.}

\item{status}{The status of the create authorizer request.}
}
\description{
Creates an authorizer.
}
\examples{

}
