% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_billing_group}
\alias{create_billing_group}
\title{Creates a billing group}
\usage{
create_billing_group(
  billingGroupName = "string",
  billingGroupProperties = list(
    billingGroupDescription = "string"
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{billingGroupName}{[required] The name you wish to give to the billing group.}

\item{billingGroupProperties}{The properties of the billing group.}

\item{tags}{Metadata which can be used to manage the billing group.}
}
\description{
Creates a billing group.
}
\examples{

}
