% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_keys_and_certificate}
\alias{create_keys_and_certificate}
\title{Creates a 2048-bit RSA key pair and issues an X}
\usage{
create_keys_and_certificate(setAsActive = NULL)
}
\arguments{
\item{setAsActive}{Specifies whether the certificate is active.}
}
\description{
Creates a 2048-bit RSA key pair and issues an X.509 certificate using the issued public key.
}
\details{
\strong{Note} This is the only time AWS IoT issues the private key for this certificate, so it is important to keep it in a secure location.
}
\section{Accepted Parameters}{
\preformatted{create_keys_and_certificate(
  setAsActive = TRUE|FALSE
)
}
}

