% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_policy_version}
\alias{create_policy_version}
\title{Creates a new version of the specified AWS IoT policy}
\usage{
create_policy_version(policyName, policyDocument, setAsDefault = NULL)
}
\arguments{
\item{policyName}{The policy name.}

\item{policyDocument}{The JSON document that describes the policy. Minimum length of 1. Maximum length of 2048, excluding whitespace.}

\item{setAsDefault}{Specifies whether the policy version is set as the default. When this parameter is true, the new policy version becomes the operative version (that is, the version that is in effect for the certificates to which the policy is attached).}
}
\description{
Creates a new version of the specified AWS IoT policy. To update a policy, create a new policy version. A managed policy can have up to five versions. If the policy has five versions, you must use DeletePolicyVersion to delete an existing version before you create a new one.
}
\details{
Optionally, you can set the new version as the policy's default version. The default version is the operative version (that is, the version that is in effect for the certificates to which the policy is attached).
}
\examples{

}
