% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_role_alias}
\alias{create_role_alias}
\title{Creates a role alias}
\usage{
create_role_alias(roleAlias, roleArn, credentialDurationSeconds = NULL)
}
\arguments{
\item{roleAlias}{The role alias that points to a role ARN. This allows you to change the role without having to update the device.}

\item{roleArn}{The role ARN.}

\item{credentialDurationSeconds}{How long (in seconds) the credentials will be valid.}
}
\description{
Creates a role alias.
}
\examples{

}
