% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_scheduled_audit}
\alias{create_scheduled_audit}
\title{Creates a scheduled audit that is run at a specified time interval}
\usage{
create_scheduled_audit(frequency, targetCheckNames, scheduledAuditName,
  dayOfMonth = NULL, dayOfWeek = NULL)
}
\arguments{
\item{frequency}{How often the scheduled audit takes place. Can be one of "DAILY", "WEEKLY", "BIWEEKLY" or "MONTHLY". The actual start time of each audit is determined by the system.}

\item{targetCheckNames}{Which checks are performed during the scheduled audit. Checks must be enabled for your account. (Use \code{DescribeAccountAuditConfiguration} to see the list of all checks including those that are enabled or \code{UpdateAccountAuditConfiguration} to select which checks are enabled.)}

\item{scheduledAuditName}{The name you want to give to the scheduled audit. (Max. 128 chars)}

\item{dayOfMonth}{The day of the month on which the scheduled audit takes place. Can be "1" through "31" or "LAST". This field is required if the "frequency" parameter is set to "MONTHLY". If days 29-31 are specified, and the month does not have that many days, the audit takes place on the "LAST" day of the month.}

\item{dayOfWeek}{The day of the week on which the scheduled audit takes place. Can be one of "SUN", "MON", "TUE", "WED", "THU", "FRI" or "SAT". This field is required if the "frequency" parameter is set to "WEEKLY" or "BIWEEKLY".}
}
\description{
Creates a scheduled audit that is run at a specified time interval.
}
\examples{

}
