% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_thing}
\alias{create_thing}
\title{Creates a thing record in the registry}
\usage{
create_thing(thingName, thingTypeName = NULL, attributePayload = NULL,
  billingGroupName = NULL)
}
\arguments{
\item{thingName}{[required] The name of the thing to create.}

\item{thingTypeName}{The name of the thing type associated with the new thing.}

\item{attributePayload}{The attribute payload, which consists of up to three name/value pairs in a JSON document. For example:

\code{{\"attributes\":{\"string1\":\"string2\"}}}}

\item{billingGroupName}{The name of the billing group the thing will be added to.}
}
\description{
Creates a thing record in the registry.
}
\details{
This is a control plane operation. See \href{http://docs.aws.amazon.com/iot/latest/developerguide/authorization.html}{Authorization} for information about authorizing control plane actions.
}
\section{Accepted Parameters}{
\preformatted{create_thing(
  thingName = "string",
  thingTypeName = "string",
  attributePayload = list(
    attributes = list(
      "string"
    ),
    merge = TRUE|FALSE
  ),
  billingGroupName = "string"
)
}
}

