% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_thing_type}
\alias{create_thing_type}
\title{Creates a new thing type}
\usage{
create_thing_type(thingTypeName, thingTypeProperties = NULL)
}
\arguments{
\item{thingTypeName}{The name of the thing type.}

\item{thingTypeProperties}{The ThingTypeProperties for the thing type to create. It contains information about the new thing type including a description, and a list of searchable thing attribute names.}
}
\description{
Creates a new thing type.
}
\examples{

}
