% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_topic_rule}
\alias{create_topic_rule}
\title{Creates a rule}
\usage{
create_topic_rule(ruleName, topicRulePayload)
}
\arguments{
\item{ruleName}{The name of the rule.}

\item{topicRulePayload}{The rule payload.}
}
\description{
Creates a rule. Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.
}
\examples{

}
