% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_account_audit_configuration}
\alias{delete_account_audit_configuration}
\title{Restores the default settings for Device Defender audits for this account}
\usage{
delete_account_audit_configuration(deleteScheduledAudits = NULL)
}
\arguments{
\item{deleteScheduledAudits}{If true, all scheduled audits are deleted.}
}
\description{
Restores the default settings for Device Defender audits for this account. Any configuration data you entered is deleted and all audit checks are reset to disabled.
}
\section{Accepted Parameters}{
\preformatted{delete_account_audit_configuration(
  deleteScheduledAudits = TRUE|FALSE
)
}
}

