% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_billing_group}
\alias{delete_billing_group}
\title{Deletes the billing group}
\usage{
delete_billing_group(billingGroupName, expectedVersion = NULL)
}
\arguments{
\item{billingGroupName}{[required] The name of the billing group.}

\item{expectedVersion}{The expected version of the billing group. If the version of the billing group does not match the expected version specified in the request, the \code{DeleteBillingGroup} request is rejected with a \code{VersionConflictException}.}
}
\description{
Deletes the billing group.
}
\section{Accepted Parameters}{
\preformatted{delete_billing_group(
  billingGroupName = "string",
  expectedVersion = 123
)
}
}

