% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_ca_certificate}
\alias{delete_ca_certificate}
\title{Deletes a registered CA certificate}
\usage{
delete_ca_certificate(certificateId)
}
\arguments{
\item{certificateId}{The ID of the certificate to delete. (The last part of the certificate ARN contains the certificate ID.)}
}
\description{
Deletes a registered CA certificate.
}
\examples{

}
