% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_dynamic_thing_group}
\alias{delete_dynamic_thing_group}
\title{Deletes a dynamic thing group}
\usage{
delete_dynamic_thing_group(
  thingGroupName = "string",
  expectedVersion = 123
)
}
\arguments{
\item{thingGroupName}{[required] The name of the dynamic thing group to delete.}

\item{expectedVersion}{The expected version of the dynamic thing group to delete.}
}
\description{
Deletes a dynamic thing group.
}
\examples{

}
