% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_job_execution}
\alias{delete_job_execution}
\title{Deletes a job execution}
\usage{
delete_job_execution(
  jobId = "string",
  thingName = "string",
  executionNumber = 123,
  force = TRUE|FALSE
)
}
\arguments{
\item{jobId}{[required] The ID of the job whose execution on a particular device will be deleted.}

\item{thingName}{[required] The name of the thing whose job execution will be deleted.}

\item{executionNumber}{[required] The ID of the job execution to be deleted. The \code{executionNumber} refers to the execution of a particular job on a particular device.

Note that once a job execution is deleted, the \code{executionNumber} may be reused by IoT, so be sure you get and use the correct value here.}

\item{force}{(Optional) When true, you can delete a job execution which is "IN\_PROGRESS". Otherwise, you can only delete a job execution which is in a terminal state ("SUCCEEDED", "FAILED", "REJECTED", "REMOVED" or "CANCELED") or an exception will occur. The default is false.

Deleting a job execution which is "IN\_PROGRESS", will cause the device to be unable to access job information or update the job execution status. Use caution and ensure that the device is able to recover to a valid state.}
}
\description{
Deletes a job execution.
}
\examples{

}
