% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_ota_update}
\alias{delete_ota_update}
\title{Delete an OTA update}
\usage{
delete_ota_update(
  otaUpdateId = "string",
  deleteStream = TRUE|FALSE,
  forceDeleteAWSJob = TRUE|FALSE
)
}
\arguments{
\item{otaUpdateId}{[required] The OTA update ID to delete.}

\item{deleteStream}{Specifies if the stream associated with an OTA update should be deleted when the OTA update is deleted.}

\item{forceDeleteAWSJob}{Specifies if the AWS Job associated with the OTA update should be deleted with the OTA update is deleted.}
}
\description{
Delete an OTA update.
}
\examples{

}
