% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_registration_code}
\alias{delete_registration_code}
\title{Deletes a CA certificate registration code}
\usage{
delete_registration_code()
}
\description{
Deletes a CA certificate registration code.
}
\examples{

}
