% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_role_alias}
\alias{delete_role_alias}
\title{Deletes a role alias}
\usage{
delete_role_alias(roleAlias)
}
\arguments{
\item{roleAlias}{[required] The role alias to delete.}
}
\description{
Deletes a role alias
}
\section{Accepted Parameters}{
\preformatted{delete_role_alias(
  roleAlias = "string"
)
}
}

