% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_scheduled_audit}
\alias{delete_scheduled_audit}
\title{Deletes a scheduled audit}
\usage{
delete_scheduled_audit(scheduledAuditName)
}
\arguments{
\item{scheduledAuditName}{[required] The name of the scheduled audit you want to delete.}
}
\description{
Deletes a scheduled audit.
}
\section{Accepted Parameters}{
\preformatted{delete_scheduled_audit(
  scheduledAuditName = "string"
)
}
}

