% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_security_profile}
\alias{delete_security_profile}
\title{Deletes a Device Defender security profile}
\usage{
delete_security_profile(securityProfileName, expectedVersion = NULL)
}
\arguments{
\item{securityProfileName}{[required] The name of the security profile to be deleted.}

\item{expectedVersion}{The expected version of the security profile. A new version is generated whenever the security profile is updated. If you specify a value that is different than the actual version, a \code{VersionConflictException} is thrown.}
}
\description{
Deletes a Device Defender security profile.
}
\section{Accepted Parameters}{
\preformatted{delete_security_profile(
  securityProfileName = "string",
  expectedVersion = 123
)
}
}

