% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_thing}
\alias{delete_thing}
\title{Deletes the specified thing}
\usage{
delete_thing(thingName, expectedVersion = NULL)
}
\arguments{
\item{thingName}{[required] The name of the thing to delete.}

\item{expectedVersion}{The expected version of the thing record in the registry. If the version of the record in the registry does not match the expected version specified in the request, the \code{DeleteThing} request is rejected with a \code{VersionConflictException}.}
}
\description{
Deletes the specified thing.
}
\section{Accepted Parameters}{
\preformatted{delete_thing(
  thingName = "string",
  expectedVersion = 123
)
}
}

