% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_thing_type}
\alias{delete_thing_type}
\title{Deletes the specified thing type}
\usage{
delete_thing_type(thingTypeName)
}
\arguments{
\item{thingTypeName}{[required] The name of the thing type.}
}
\description{
Deletes the specified thing type. You cannot delete a thing type if it has things associated with it. To delete a thing type, first mark it as deprecated by calling DeprecateThingType, then remove any associated things by calling UpdateThing to change the thing type on any associated thing, and finally use DeleteThingType to delete the thing type.
}
\section{Accepted Parameters}{
\preformatted{delete_thing_type(
  thingTypeName = "string"
)
}
}

