% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_topic_rule}
\alias{delete_topic_rule}
\title{Deletes the rule}
\usage{
delete_topic_rule(
  ruleName = "string"
)
}
\arguments{
\item{ruleName}{[required] The name of the rule.}
}
\description{
Deletes the rule.
}
\examples{

}
