% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_v2_logging_level}
\alias{delete_v2_logging_level}
\title{Deletes a logging level}
\usage{
delete_v2_logging_level(targetType, targetName)
}
\arguments{
\item{targetType}{The type of resource for which you are configuring logging. Must be \code{THING_Group}.}

\item{targetName}{The name of the resource for which you are configuring logging.}
}
\description{
Deletes a logging level.
}
\examples{

}
