% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{deprecate_thing_type}
\alias{deprecate_thing_type}
\title{Deprecates a thing type}
\usage{
deprecate_thing_type(thingTypeName, undoDeprecate = NULL)
}
\arguments{
\item{thingTypeName}{The name of the thing type to deprecate.}

\item{undoDeprecate}{Whether to undeprecate a deprecated thing type. If \strong{true}, the thing type will not be deprecated anymore and you can associate it with things.}
}
\description{
Deprecates a thing type. You can not associate new things with deprecated thing type.
}
\examples{

}
