% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_account_audit_configuration}
\alias{describe_account_audit_configuration}
\title{Gets information about the Device Defender audit settings for this account}
\usage{
describe_account_audit_configuration()
}
\description{
Gets information about the Device Defender audit settings for this account. Settings include how audit notifications are sent and which audit checks are enabled or disabled.
}
\section{Accepted Parameters}{
\preformatted{describe_account_audit_configuration()
}
}

