% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_audit_task}
\alias{describe_audit_task}
\title{Gets information about a Device Defender audit}
\usage{
describe_audit_task(taskId)
}
\arguments{
\item{taskId}{[required] The ID of the audit whose information you want to get.}
}
\description{
Gets information about a Device Defender audit.
}
\section{Accepted Parameters}{
\preformatted{describe_audit_task(
  taskId = "string"
)
}
}

