% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_ca_certificate}
\alias{describe_ca_certificate}
\title{Describes a registered CA certificate}
\usage{
describe_ca_certificate(certificateId)
}
\arguments{
\item{certificateId}{[required] The CA certificate identifier.}
}
\description{
Describes a registered CA certificate.
}
\section{Accepted Parameters}{
\preformatted{describe_ca_certificate(
  certificateId = "string"
)
}
}

