% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_certificate}
\alias{describe_certificate}
\title{Gets information about the specified certificate}
\usage{
describe_certificate(certificateId)
}
\arguments{
\item{certificateId}{[required] The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)}
}
\description{
Gets information about the specified certificate.
}
\section{Accepted Parameters}{
\preformatted{describe_certificate(
  certificateId = "string"
)
}
}

