% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_default_authorizer}
\alias{describe_default_authorizer}
\title{Describes the default authorizer}
\usage{
describe_default_authorizer()
}
\description{
Describes the default authorizer.
}
\section{Accepted Parameters}{
\preformatted{describe_default_authorizer()
}
}

