% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_endpoint}
\alias{describe_endpoint}
\title{Returns a unique endpoint specific to the AWS account making the call}
\usage{
describe_endpoint(endpointType = NULL)
}
\arguments{
\item{endpointType}{The endpoint type. Valid endpoint types include:
\itemize{
\item \code{iot:Data} - Returns a VeriSign signed data endpoint.
\item \code{iot:Data-ATS} - Returns an ATS signed data endpoint.
\item \code{iot:CredentialProvider} - Returns an AWS IoT credentials provider API endpoint.
\item \code{iot:Jobs} - Returns an AWS IoT device management Jobs API endpoint.
}}
}
\description{
Returns a unique endpoint specific to the AWS account making the call.
}
\section{Accepted Parameters}{
\preformatted{describe_endpoint(
  endpointType = "string"
)
}
}

