% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_event_configurations}
\alias{describe_event_configurations}
\title{Describes event configurations}
\usage{
describe_event_configurations()
}
\description{
Describes event configurations.
}
\examples{

}
