% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_job}
\alias{describe_job}
\title{Describes a job}
\usage{
describe_job(jobId)
}
\arguments{
\item{jobId}{The unique identifier you assigned to this job when it was created.}
}
\description{
Describes a job.
}
\examples{

}
