% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_job_execution}
\alias{describe_job_execution}
\title{Describes a job execution}
\usage{
describe_job_execution(jobId, thingName, executionNumber = NULL)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}

\item{thingName}{[required] The name of the thing on which the job execution is running.}

\item{executionNumber}{A string (consisting of the digits "0" through "9" which is used to specify a particular job execution on a particular device.}
}
\description{
Describes a job execution.
}
\section{Accepted Parameters}{
\preformatted{describe_job_execution(
  jobId = "string",
  thingName = "string",
  executionNumber = 123
)
}
}

