% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_role_alias}
\alias{describe_role_alias}
\title{Describes a role alias}
\usage{
describe_role_alias(
  roleAlias = "string"
)
}
\arguments{
\item{roleAlias}{[required] The role alias to describe.}
}
\description{
Describes a role alias.
}
\examples{

}
