% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_scheduled_audit}
\alias{describe_scheduled_audit}
\title{Gets information about a scheduled audit}
\usage{
describe_scheduled_audit(scheduledAuditName)
}
\arguments{
\item{scheduledAuditName}{The name of the scheduled audit whose information you want to get.}
}
\description{
Gets information about a scheduled audit.
}
\examples{

}
