% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_security_profile}
\alias{describe_security_profile}
\title{Gets information about a Device Defender security profile}
\usage{
describe_security_profile(securityProfileName)
}
\arguments{
\item{securityProfileName}{[required] The name of the security profile whose information you want to get.}
}
\description{
Gets information about a Device Defender security profile.
}
\section{Accepted Parameters}{
\preformatted{describe_security_profile(
  securityProfileName = "string"
)
}
}

