% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_stream}
\alias{describe_stream}
\title{Gets information about a stream}
\usage{
describe_stream(streamId)
}
\arguments{
\item{streamId}{The stream ID.}
}
\description{
Gets information about a stream.
}
\examples{

}
