% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_thing}
\alias{describe_thing}
\title{Gets information about the specified thing}
\usage{
describe_thing(thingName)
}
\arguments{
\item{thingName}{[required] The name of the thing.}
}
\description{
Gets information about the specified thing.
}
\section{Accepted Parameters}{
\preformatted{describe_thing(
  thingName = "string"
)
}
}

