% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_thing_group}
\alias{describe_thing_group}
\title{Describe a thing group}
\usage{
describe_thing_group(
  thingGroupName = "string"
)
}
\arguments{
\item{thingGroupName}{[required] The name of the thing group.}
}
\description{
Describe a thing group.
}
\examples{

}
