% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_thing_registration_task}
\alias{describe_thing_registration_task}
\title{Describes a bulk thing provisioning task}
\usage{
describe_thing_registration_task(taskId)
}
\arguments{
\item{taskId}{The task ID.}
}
\description{
Describes a bulk thing provisioning task.
}
\examples{

}
