% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{detach_principal_policy}
\alias{detach_principal_policy}
\title{Removes the specified policy from the specified certificate}
\usage{
detach_principal_policy(policyName, principal)
}
\arguments{
\item{policyName}{The name of the policy to detach.}

\item{principal}{The principal.

If the principal is a certificate, specify the certificate ARN. If the principal is an Amazon Cognito identity, specify the identity ID.}
}
\description{
Removes the specified policy from the specified certificate.
}
\details{
\strong{Note:} This API is deprecated. Please use DetachPolicy instead.
}
\examples{

}
