% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{disable_topic_rule}
\alias{disable_topic_rule}
\title{Disables the rule}
\usage{
disable_topic_rule(ruleName)
}
\arguments{
\item{ruleName}{The name of the rule to disable.}
}
\description{
Disables the rule.
}
\examples{

}
