% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{enable_topic_rule}
\alias{enable_topic_rule}
\title{Enables the rule}
\usage{
enable_topic_rule(
  ruleName = "string"
)
}
\arguments{
\item{ruleName}{[required] The name of the topic rule to enable.}
}
\description{
Enables the rule.
}
\examples{

}
