% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_effective_policies}
\alias{get_effective_policies}
\title{Gets a list of the policies that have an effect on the authorization behavior of the specified device when it connects to the AWS IoT device gateway}
\usage{
get_effective_policies(
  principal = "string",
  cognitoIdentityPoolId = "string",
  thingName = "string"
)
}
\arguments{
\item{principal}{The principal.}

\item{cognitoIdentityPoolId}{The Cognito identity pool ID.}

\item{thingName}{The thing name.}
}
\description{
Gets a list of the policies that have an effect on the authorization behavior of the specified device when it connects to the AWS IoT device gateway.
}
\examples{

}
