% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_indexing_configuration}
\alias{get_indexing_configuration}
\title{Gets the search configuration}
\usage{
get_indexing_configuration()
}
\description{
Gets the search configuration.
}
\section{Accepted Parameters}{
\preformatted{get_indexing_configuration()
}
}

