% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_job_document}
\alias{get_job_document}
\title{Gets a job document}
\usage{
get_job_document(
  jobId = "string"
)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}
}
\description{
Gets a job document.
}
\examples{

}
