% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_logging_options}
\alias{get_logging_options}
\title{Gets the logging options}
\usage{
get_logging_options()
}
\description{
Gets the logging options.
}
\details{
NOTE: use of this command is not recommended. Use \code{GetV2LoggingOptions} instead.
}
\section{Accepted Parameters}{
\preformatted{get_logging_options()
}
}

