% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_policy}
\alias{get_policy}
\title{Gets information about the specified policy with the policy document of the default version}
\usage{
get_policy(policyName)
}
\arguments{
\item{policyName}{[required] The name of the policy.}
}
\description{
Gets information about the specified policy with the policy document of the default version.
}
\section{Accepted Parameters}{
\preformatted{get_policy(
  policyName = "string"
)
}
}

