% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_policy_version}
\alias{get_policy_version}
\title{Gets information about the specified policy version}
\usage{
get_policy_version(policyName, policyVersionId)
}
\arguments{
\item{policyName}{[required] The name of the policy.}

\item{policyVersionId}{[required] The policy version ID.}
}
\description{
Gets information about the specified policy version.
}
\section{Accepted Parameters}{
\preformatted{get_policy_version(
  policyName = "string",
  policyVersionId = "string"
)
}
}

