% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_registration_code}
\alias{get_registration_code}
\title{Gets a registration code used to register a CA certificate with AWS IoT}
\usage{
get_registration_code()
}
\description{
Gets a registration code used to register a CA certificate with AWS IoT.
}
\section{Accepted Parameters}{
\preformatted{get_registration_code()
}
}

