% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_topic_rule}
\alias{get_topic_rule}
\title{Gets information about the rule}
\usage{
get_topic_rule(
  ruleName = "string"
)
}
\arguments{
\item{ruleName}{[required] The name of the rule.}
}
\description{
Gets information about the rule.
}
\examples{

}
