% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_v2_logging_options}
\alias{get_v2_logging_options}
\title{Gets the fine grained logging options}
\usage{
get_v2_logging_options()
}
\description{
Gets the fine grained logging options.
}
\examples{

}
