% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_active_violations}
\alias{list_active_violations}
\title{Lists the active violations for a given Device Defender security profile}
\usage{
list_active_violations(thingName = NULL, securityProfileName = NULL,
  nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{thingName}{The name of the thing whose active violations are listed.}

\item{securityProfileName}{The name of the Device Defender security profile for which violations are listed.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the active violations for a given Device Defender security profile.
}
\section{Accepted Parameters}{
\preformatted{list_active_violations(
  thingName = "string",
  securityProfileName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

