% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_attached_policies}
\alias{list_attached_policies}
\title{Lists the policies attached to the specified thing group}
\usage{
list_attached_policies(target, recursive = NULL, marker = NULL,
  pageSize = NULL)
}
\arguments{
\item{target}{[required] The group for which the policies will be listed.}

\item{recursive}{When true, recursively list attached policies.}

\item{marker}{The token to retrieve the next set of results.}

\item{pageSize}{The maximum number of results to be returned per request.}
}
\description{
Lists the policies attached to the specified thing group.
}
\section{Accepted Parameters}{
\preformatted{list_attached_policies(
  target = "string",
  recursive = TRUE|FALSE,
  marker = "string",
  pageSize = 123
)
}
}

