% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_audit_findings}
\alias{list_audit_findings}
\title{Lists the findings (results) of a Device Defender audit or of the audits performed during a specified time period}
\usage{
list_audit_findings(taskId = NULL, checkName = NULL,
  resourceIdentifier = NULL, maxResults = NULL, nextToken = NULL,
  startTime = NULL, endTime = NULL)
}
\arguments{
\item{taskId}{A filter to limit results to the audit with the specified ID. You must specify either the taskId or the startTime and endTime, but not both.}

\item{checkName}{A filter to limit results to the findings for the specified audit check.}

\item{resourceIdentifier}{Information identifying the non-compliant resource.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}

\item{nextToken}{The token for the next set of results.}

\item{startTime}{A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.}

\item{endTime}{A filter to limit results to those found before the specified time. You must specify either the startTime and endTime or the taskId, but not both.}
}
\description{
Lists the findings (results) of a Device Defender audit or of the audits performed during a specified time period. (Findings are retained for 180 days.)
}
\examples{

}
