% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_billing_groups}
\alias{list_billing_groups}
\title{Lists the billing groups you have created}
\usage{
list_billing_groups(
  nextToken = "string",
  maxResults = 123,
  namePrefixFilter = "string"
)
}
\arguments{
\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return per request.}

\item{namePrefixFilter}{Limit the results to billing groups whose names have the given prefix.}
}
\description{
Lists the billing groups you have created.
}
\examples{

}
