% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_certificates}
\alias{list_certificates}
\title{Lists the certificates registered in your AWS account}
\usage{
list_certificates(pageSize = NULL, marker = NULL, ascendingOrder = NULL)
}
\arguments{
\item{pageSize}{The result page size.}

\item{marker}{The marker for the next set of results.}

\item{ascendingOrder}{Specifies the order for results. If True, the results are returned in ascending order, based on the creation date.}
}
\description{
Lists the certificates registered in your AWS account.
}
\details{
The results are paginated with a default page size of 25. You can use the returned marker to retrieve additional results.
}
\examples{

}
