% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_certificates_by_ca}
\alias{list_certificates_by_ca}
\title{List the device certificates signed by the specified CA certificate}
\usage{
list_certificates_by_ca(caCertificateId, pageSize = NULL, marker = NULL,
  ascendingOrder = NULL)
}
\arguments{
\item{caCertificateId}{The ID of the CA certificate. This operation will list all registered device certificate that were signed by this CA certificate.}

\item{pageSize}{The result page size.}

\item{marker}{The marker for the next set of results.}

\item{ascendingOrder}{Specifies the order for results. If True, the results are returned in ascending order, based on the creation date.}
}
\description{
List the device certificates signed by the specified CA certificate.
}
\examples{

}
