% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_indices}
\alias{list_indices}
\title{Lists the search indices}
\usage{
list_indices(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The token used to get the next set of results, or \strong{null} if there are no additional results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the search indices.
}
\section{Accepted Parameters}{
\preformatted{list_indices(
  nextToken = "string",
  maxResults = 123
)
}
}

