% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_job_executions_for_job}
\alias{list_job_executions_for_job}
\title{Lists the job executions for a job}
\usage{
list_job_executions_for_job(jobId, status = NULL, maxResults = NULL,
  nextToken = NULL)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}

\item{status}{The status of the job.}

\item{maxResults}{The maximum number of results to be returned per request.}

\item{nextToken}{The token to retrieve the next set of results.}
}
\description{
Lists the job executions for a job.
}
\section{Accepted Parameters}{
\preformatted{list_job_executions_for_job(
  jobId = "string",
  status = "QUEUED"|"IN_PROGRESS"|"SUCCEEDED"|"FAILED"|"TIMED_OUT"|"REJECTED"|"REMOVED"|"CANCELED",
  maxResults = 123,
  nextToken = "string"
)
}
}

