% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_job_executions_for_thing}
\alias{list_job_executions_for_thing}
\title{Lists the job executions for the specified thing}
\usage{
list_job_executions_for_thing(thingName, status = NULL, maxResults = NULL,
  nextToken = NULL)
}
\arguments{
\item{thingName}{The thing name.}

\item{status}{An optional filter that lets you search for jobs that have the specified status.}

\item{maxResults}{The maximum number of results to be returned per request.}

\item{nextToken}{The token to retrieve the next set of results.}
}
\description{
Lists the job executions for the specified thing.
}
\examples{

}
