% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_policies}
\alias{list_policies}
\title{Lists your policies}
\usage{
list_policies(
  marker = "string",
  pageSize = 123,
  ascendingOrder = TRUE|FALSE
)
}
\arguments{
\item{marker}{The marker for the next set of results.}

\item{pageSize}{The result page size.}

\item{ascendingOrder}{Specifies the order for results. If true, the results are returned in ascending creation order.}
}
\description{
Lists your policies.
}
\examples{

}
