% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_policy_principals}
\alias{list_policy_principals}
\title{Lists the principals associated with the specified policy}
\usage{
list_policy_principals(policyName, marker = NULL, pageSize = NULL,
  ascendingOrder = NULL)
}
\arguments{
\item{policyName}{The policy name.}

\item{marker}{The marker for the next set of results.}

\item{pageSize}{The result page size.}

\item{ascendingOrder}{Specifies the order for results. If true, the results are returned in ascending creation order.}
}
\description{
Lists the principals associated with the specified policy.
}
\details{
\strong{Note:} This API is deprecated. Please use ListTargetsForPolicy instead.
}
\examples{

}
