% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_policy_versions}
\alias{list_policy_versions}
\title{Lists the versions of the specified policy and identifies the default version}
\usage{
list_policy_versions(
  policyName = "string"
)
}
\arguments{
\item{policyName}{[required] The policy name.}
}
\description{
Lists the versions of the specified policy and identifies the default version.
}
\examples{

}
