% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_principal_policies}
\alias{list_principal_policies}
\title{Lists the policies attached to the specified principal}
\usage{
list_principal_policies(principal, marker = NULL, pageSize = NULL,
  ascendingOrder = NULL)
}
\arguments{
\item{principal}{[required] The principal.}

\item{marker}{The marker for the next set of results.}

\item{pageSize}{The result page size.}

\item{ascendingOrder}{Specifies the order for results. If true, results are returned in ascending creation order.}
}
\description{
Lists the policies attached to the specified principal. If you use an Cognito identity, the ID must be in \href{http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_GetCredentialsForIdentity.html#API_GetCredentialsForIdentity_RequestSyntax}{AmazonCognito Identity format}.
}
\details{
\strong{Note:} This API is deprecated. Please use ListAttachedPolicies instead.
}
\section{Accepted Parameters}{
\preformatted{list_principal_policies(
  principal = "string",
  marker = "string",
  pageSize = 123,
  ascendingOrder = TRUE|FALSE
)
}
}

