% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_principal_things}
\alias{list_principal_things}
\title{Lists the things associated with the specified principal}
\usage{
list_principal_things(principal, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{principal}{The principal.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return in this operation.}
}
\description{
Lists the things associated with the specified principal.
}
\examples{

}
