% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_role_aliases}
\alias{list_role_aliases}
\title{Lists the role aliases registered in your account}
\usage{
list_role_aliases(
  pageSize = 123,
  marker = "string",
  ascendingOrder = TRUE|FALSE
)
}
\arguments{
\item{pageSize}{The maximum number of results to return at one time.}

\item{marker}{A marker used to get the next set of results.}

\item{ascendingOrder}{Return the list of role aliases in ascending alphabetical order.}
}
\description{
Lists the role aliases registered in your account.
}
\examples{

}
