% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_scheduled_audits}
\alias{list_scheduled_audits}
\title{Lists all of your scheduled audits}
\usage{
list_scheduled_audits(
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}
}
\description{
Lists all of your scheduled audits.
}
\examples{

}
