% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_security_profiles}
\alias{list_security_profiles}
\title{Lists the Device Defender security profiles you have created}
\usage{
list_security_profiles(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the Device Defender security profiles you have created. You can use filters to list only those security profiles associated with a thing group or only those associated with your account.
}
\section{Accepted Parameters}{
\preformatted{list_security_profiles(
  nextToken = "string",
  maxResults = 123
)
}
}

