% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_security_profiles_for_target}
\alias{list_security_profiles_for_target}
\title{Lists the Device Defender security profiles attached to a target (thing group)}
\usage{
list_security_profiles_for_target(securityProfileTargetArn, nextToken = NULL,
  maxResults = NULL, recursive = NULL)
}
\arguments{
\item{securityProfileTargetArn}{The ARN of the target (thing group) whose attached security profiles you want to get.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{recursive}{If true, return child groups as well.}
}
\description{
Lists the Device Defender security profiles attached to a target (thing group).
}
\examples{

}
