% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_streams}
\alias{list_streams}
\title{Lists all of the streams in your AWS account}
\usage{
list_streams(maxResults = NULL, nextToken = NULL, ascendingOrder = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return at a time.}

\item{nextToken}{A token used to get the next set of results.}

\item{ascendingOrder}{Set to true to return the list of streams in ascending order.}
}
\description{
Lists all of the streams in your AWS account.
}
\examples{

}
