% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists the tags (metadata) you have assigned to the resource}
\usage{
list_tags_for_resource(resourceArn, nextToken = NULL)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{nextToken}{The token to retrieve the next set of results.}
}
\description{
Lists the tags (metadata) you have assigned to the resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  resourceArn = "string",
  nextToken = "string"
)
}
}

