% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_targets_for_policy}
\alias{list_targets_for_policy}
\title{List targets for the specified policy}
\usage{
list_targets_for_policy(policyName, marker = NULL, pageSize = NULL)
}
\arguments{
\item{policyName}{[required] The policy name.}

\item{marker}{A marker used to get the next set of results.}

\item{pageSize}{The maximum number of results to return at one time.}
}
\description{
List targets for the specified policy.
}
\section{Accepted Parameters}{
\preformatted{list_targets_for_policy(
  policyName = "string",
  marker = "string",
  pageSize = 123
)
}
}

