% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_targets_for_security_profile}
\alias{list_targets_for_security_profile}
\title{Lists the targets (thing groups) associated with a given Device Defender security profile}
\usage{
list_targets_for_security_profile(securityProfileName, nextToken = NULL,
  maxResults = NULL)
}
\arguments{
\item{securityProfileName}{[required] The security profile.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the targets (thing groups) associated with a given Device Defender security profile.
}
\section{Accepted Parameters}{
\preformatted{list_targets_for_security_profile(
  securityProfileName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

