% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_thing_groups}
\alias{list_thing_groups}
\title{List the thing groups in your account}
\usage{
list_thing_groups(
  nextToken = "string",
  maxResults = 123,
  parentGroup = "string",
  namePrefixFilter = "string",
  recursive = TRUE|FALSE
)
}
\arguments{
\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{parentGroup}{A filter that limits the results to those with the specified parent group.}

\item{namePrefixFilter}{A filter that limits the results to those with the specified name prefix.}

\item{recursive}{If true, return child groups as well.}
}
\description{
List the thing groups in your account.
}
\examples{

}
