% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_thing_groups_for_thing}
\alias{list_thing_groups_for_thing}
\title{List the thing groups to which the specified thing belongs}
\usage{
list_thing_groups_for_thing(thingName, nextToken = NULL,
  maxResults = NULL)
}
\arguments{
\item{thingName}{[required] The thing name.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
List the thing groups to which the specified thing belongs.
}
\section{Accepted Parameters}{
\preformatted{list_thing_groups_for_thing(
  thingName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

