% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_thing_principals}
\alias{list_thing_principals}
\title{Lists the principals associated with the specified thing}
\usage{
list_thing_principals(thingName)
}
\arguments{
\item{thingName}{[required] The name of the thing.}
}
\description{
Lists the principals associated with the specified thing.
}
\section{Accepted Parameters}{
\preformatted{list_thing_principals(
  thingName = "string"
)
}
}

