% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_thing_registration_task_reports}
\alias{list_thing_registration_task_reports}
\title{Information about the thing registration tasks}
\usage{
list_thing_registration_task_reports(taskId, reportType, nextToken = NULL,
  maxResults = NULL)
}
\arguments{
\item{taskId}{The id of the task.}

\item{reportType}{The type of task report.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return per request.}
}
\description{
Information about the thing registration tasks.
}
\examples{

}
