% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_thing_types}
\alias{list_thing_types}
\title{Lists the existing thing types}
\usage{
list_thing_types(
  nextToken = "string",
  maxResults = 123,
  thingTypeName = "string"
)
}
\arguments{
\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return in this operation.}

\item{thingTypeName}{The name of the thing type.}
}
\description{
Lists the existing thing types.
}
\examples{

}
