% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_topic_rules}
\alias{list_topic_rules}
\title{Lists the rules for the specific topic}
\usage{
list_topic_rules(topic = NULL, maxResults = NULL, nextToken = NULL,
  ruleDisabled = NULL)
}
\arguments{
\item{topic}{The topic.}

\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A token used to retrieve the next value.}

\item{ruleDisabled}{Specifies whether the rule is disabled.}
}
\description{
Lists the rules for the specific topic.
}
\examples{

}
