% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_v2_logging_levels}
\alias{list_v2_logging_levels}
\title{Lists logging levels}
\usage{
list_v2_logging_levels(targetType = NULL, nextToken = NULL,
  maxResults = NULL)
}
\arguments{
\item{targetType}{The type of resource for which you are configuring logging. Must be \code{THING_Group}.}

\item{nextToken}{The token used to get the next set of results, or \strong{null} if there are no additional results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists logging levels.
}
\examples{

}
