% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_violation_events}
\alias{list_violation_events}
\title{Lists the Device Defender security profile violations discovered during the given time period}
\usage{
list_violation_events(startTime, endTime, thingName = NULL,
  securityProfileName = NULL, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{startTime}{The start time for the alerts to be listed.}

\item{endTime}{The end time for the alerts to be listed.}

\item{thingName}{A filter to limit results to those alerts caused by the specified thing.}

\item{securityProfileName}{A filter to limit results to those alerts generated by the specified security profile.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the Device Defender security profile violations discovered during the given time period. You can use filters to limit the results to those alerts issued for a particular security profile, behavior or thing (device).
}
\examples{

}
