% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_package.R
\docType{package}
\name{paws.iot-package}
\alias{paws.iot}
\alias{paws.iot-package}
\title{paws.iot: AWS IoT}
\description{
AWS IoT provides secure, bi-directional communication between
Internet-connected devices (such as sensors, actuators, embedded
devices, or smart appliances) and the AWS cloud. You can discover your
custom IoT-Data endpoint to communicate with, configure rules for data
processing and integration with other services, organize resources
associated with each device (Registry), configure logging, and create
and manage policies and credentials to authenticate devices.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
