% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{register_certificate}
\alias{register_certificate}
\title{Registers a device certificate with AWS IoT}
\usage{
register_certificate(certificatePem, caCertificatePem = NULL,
  setAsActive = NULL, status = NULL)
}
\arguments{
\item{certificatePem}{The certificate data, in PEM format.}

\item{caCertificatePem}{The CA certificate used to sign the device certificate being registered.}

\item{setAsActive}{A boolean value that specifies if the CA certificate is set to active.}

\item{status}{The status of the register certificate request.}
}
\description{
Registers a device certificate with AWS IoT. If you have more than one CA certificate that has the same subject field, you must specify the CA certificate that was used to sign the device certificate being registered.
}
\examples{

}
