% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{register_thing}
\alias{register_thing}
\title{Provisions a thing}
\usage{
register_thing(templateBody, parameters = NULL)
}
\arguments{
\item{templateBody}{The provisioning template. See \href{http://docs.aws.amazon.com/iot/latest/developerguide/programmatic-provisioning.html}{Programmatic Provisioning} for more information.}

\item{parameters}{The parameters for provisioning a thing. See \href{http://docs.aws.amazon.com/iot/latest/developerguide/programmatic-provisioning.html}{Programmatic Provisioning} for more information.}
}
\description{
Provisions a thing.
}
\examples{

}
