% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{remove_thing_from_thing_group}
\alias{remove_thing_from_thing_group}
\title{Remove the specified thing from the specified group}
\usage{
remove_thing_from_thing_group(
  thingGroupName = "string",
  thingGroupArn = "string",
  thingName = "string",
  thingArn = "string"
)
}
\arguments{
\item{thingGroupName}{The group name.}

\item{thingGroupArn}{The group ARN.}

\item{thingName}{The name of the thing to remove from the group.}

\item{thingArn}{The ARN of the thing to remove from the group.}
}
\description{
Remove the specified thing from the specified group.
}
\examples{

}
