% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{replace_topic_rule}
\alias{replace_topic_rule}
\title{Replaces the rule}
\usage{
replace_topic_rule(ruleName, topicRulePayload)
}
\arguments{
\item{ruleName}{The name of the rule.}

\item{topicRulePayload}{The rule payload.}
}
\description{
Replaces the rule. You must specify all parameters for the new rule. Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.
}
\examples{

}
