% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{set_default_authorizer}
\alias{set_default_authorizer}
\title{Sets the default authorizer}
\usage{
set_default_authorizer(authorizerName)
}
\arguments{
\item{authorizerName}{The authorizer name.}
}
\description{
Sets the default authorizer. This will be used if a websocket connection is made without specifying an authorizer.
}
\examples{

}
