% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{set_v2_logging_level}
\alias{set_v2_logging_level}
\title{Sets the logging level}
\usage{
set_v2_logging_level(logTarget, logLevel)
}
\arguments{
\item{logTarget}{The log target.}

\item{logLevel}{The log level.}
}
\description{
Sets the logging level.
}
\examples{

}
