% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{set_v2_logging_options}
\alias{set_v2_logging_options}
\title{Sets the logging options for the V2 logging service}
\usage{
set_v2_logging_options(
  roleArn = "string",
  defaultLogLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED",
  disableAllLogs = TRUE|FALSE
)
}
\arguments{
\item{roleArn}{The ARN of the role that allows IoT to write to Cloudwatch logs.}

\item{defaultLogLevel}{The default logging level.}

\item{disableAllLogs}{If true all logs are disabled. The default is false.}
}
\description{
Sets the logging options for the V2 logging service.
}
\examples{

}
