% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{start_thing_registration_task}
\alias{start_thing_registration_task}
\title{Creates a bulk thing provisioning task}
\usage{
start_thing_registration_task(templateBody, inputFileBucket, inputFileKey,
  roleArn)
}
\arguments{
\item{templateBody}{The provisioning template.}

\item{inputFileBucket}{The S3 bucket that contains the input file.}

\item{inputFileKey}{The name of input file within the S3 bucket. This file contains a newline delimited JSON file. Each line contains the parameter values to provision one device (thing).}

\item{roleArn}{The IAM role ARN that grants permission the input file.}
}
\description{
Creates a bulk thing provisioning task.
}
\examples{

}
