% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{stop_thing_registration_task}
\alias{stop_thing_registration_task}
\title{Cancels a bulk thing provisioning task}
\usage{
stop_thing_registration_task(taskId)
}
\arguments{
\item{taskId}{[required] The bulk thing provisioning task ID.}
}
\description{
Cancels a bulk thing provisioning task.
}
\section{Accepted Parameters}{
\preformatted{stop_thing_registration_task(
  taskId = "string"
)
}
}

