% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Adds to or modifies the tags of the given resource}
\usage{
tag_resource(
  resourceArn = "string",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tags}{[required] The new or modified tags for the resource.}
}
\description{
Adds to or modifies the tags of the given resource. Tags are metadata which can be used to manage a resource.
}
\examples{

}
