% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{test_invoke_authorizer}
\alias{test_invoke_authorizer}
\title{Tests a custom authorization behavior by invoking a specified custom authorizer}
\usage{
test_invoke_authorizer(authorizerName, token, tokenSignature)
}
\arguments{
\item{authorizerName}{[required] The custom authorizer name.}

\item{token}{[required] The token returned by your custom authentication service.}

\item{tokenSignature}{[required] The signature made with the token and your custom authentication service's private key.}
}
\description{
Tests a custom authorization behavior by invoking a specified custom authorizer. Use this to test and debug the custom authorization behavior of devices that connect to the AWS IoT device gateway.
}
\section{Accepted Parameters}{
\preformatted{test_invoke_authorizer(
  authorizerName = "string",
  token = "string",
  tokenSignature = "string"
)
}
}

