% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes the given tags (metadata) from the resource}
\usage{
untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tagKeys}{[required] A list of the keys of the tags to be removed from the resource.}
}
\description{
Removes the given tags (metadata) from the resource.
}
\examples{

}
