% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_authorizer}
\alias{update_authorizer}
\title{Updates an authorizer}
\usage{
update_authorizer(authorizerName, authorizerFunctionArn = NULL,
  tokenKeyName = NULL, tokenSigningPublicKeys = NULL, status = NULL)
}
\arguments{
\item{authorizerName}{The authorizer name.}

\item{authorizerFunctionArn}{The ARN of the authorizer's Lambda function.}

\item{tokenKeyName}{The key used to extract the token from the HTTP headers.}

\item{tokenSigningPublicKeys}{The public keys used to verify the token signature.}

\item{status}{The status of the update authorizer request.}
}
\description{
Updates an authorizer.
}
\examples{

}
