% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_billing_group}
\alias{update_billing_group}
\title{Updates information about the billing group}
\usage{
update_billing_group(billingGroupName, billingGroupProperties,
  expectedVersion = NULL)
}
\arguments{
\item{billingGroupName}{[required] The name of the billing group.}

\item{billingGroupProperties}{[required] The properties of the billing group.}

\item{expectedVersion}{The expected version of the billing group. If the version of the billing group does not match the expected version specified in the request, the \code{UpdateBillingGroup} request is rejected with a \code{VersionConflictException}.}
}
\description{
Updates information about the billing group.
}
\section{Accepted Parameters}{
\preformatted{update_billing_group(
  billingGroupName = "string",
  billingGroupProperties = list(
    billingGroupDescription = "string"
  ),
  expectedVersion = 123
)
}
}

