% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_ca_certificate}
\alias{update_ca_certificate}
\title{Updates a registered CA certificate}
\usage{
update_ca_certificate(
  certificateId = "string",
  newStatus = "ACTIVE"|"INACTIVE",
  newAutoRegistrationStatus = "ENABLE"|"DISABLE",
  registrationConfig = list(
    templateBody = "string",
    roleArn = "string"
  ),
  removeAutoRegistration = TRUE|FALSE
)
}
\arguments{
\item{certificateId}{[required] The CA certificate identifier.}

\item{newStatus}{The updated status of the CA certificate.

\strong{Note:} The status value REGISTER\_INACTIVE is deprecated and should not be used.}

\item{newAutoRegistrationStatus}{The new value for the auto registration status. Valid values are: "ENABLE" or "DISABLE".}

\item{registrationConfig}{Information about the registration configuration.}

\item{removeAutoRegistration}{If true, remove auto registration.}
}
\description{
Updates a registered CA certificate.
}
\examples{

}
