% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_indexing_configuration}
\alias{update_indexing_configuration}
\title{Updates the search configuration}
\usage{
update_indexing_configuration(thingIndexingConfiguration = NULL,
  thingGroupIndexingConfiguration = NULL)
}
\arguments{
\item{thingIndexingConfiguration}{Thing indexing configuration.}

\item{thingGroupIndexingConfiguration}{Thing group indexing configuration.}
}
\description{
Updates the search configuration.
}
\section{Accepted Parameters}{
\preformatted{update_indexing_configuration(
  thingIndexingConfiguration = list(
    thingIndexingMode = "OFF"|"REGISTRY"|"REGISTRY_AND_SHADOW",
    thingConnectivityIndexingMode = "OFF"|"STATUS"
  ),
  thingGroupIndexingConfiguration = list(
    thingGroupIndexingMode = "OFF"|"ON"
  )
)
}
}

