% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_security_profile}
\alias{update_security_profile}
\title{Updates a Device Defender security profile}
\usage{
update_security_profile(securityProfileName,
  securityProfileDescription = NULL, behaviors = NULL,
  alertTargets = NULL, expectedVersion = NULL)
}
\arguments{
\item{securityProfileName}{The name of the security profile you want to update.}

\item{securityProfileDescription}{A description of the security profile.}

\item{behaviors}{Specifies the behaviors that, when violated by a device (thing), cause an alert.}

\item{alertTargets}{Where the alerts are sent. (Alerts are always sent to the console.)}

\item{expectedVersion}{The expected version of the security profile. A new version is generated whenever the security profile is updated. If you specify a value that is different than the actual version, a \code{VersionConflictException} is thrown.}
}
\description{
Updates a Device Defender security profile.
}
\examples{

}
