% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_stream}
\alias{update_stream}
\title{Updates an existing stream}
\usage{
update_stream(streamId, description = NULL, files = NULL, roleArn = NULL)
}
\arguments{
\item{streamId}{The stream ID.}

\item{description}{The description of the stream.}

\item{files}{The files associated with the stream.}

\item{roleArn}{An IAM role that allows the IoT service principal assumes to access your S3 files.}
}
\description{
Updates an existing stream. The stream version will be incremented by one.
}
\examples{

}
