% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_thing_group}
\alias{update_thing_group}
\title{Update a thing group}
\usage{
update_thing_group(thingGroupName, thingGroupProperties,
  expectedVersion = NULL)
}
\arguments{
\item{thingGroupName}{The thing group to update.}

\item{thingGroupProperties}{The thing group properties.}

\item{expectedVersion}{The expected version of the thing group. If this does not match the version of the thing group being updated, the update will fail.}
}
\description{
Update a thing group.
}
\examples{

}
