% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{validate_security_profile_behaviors}
\alias{validate_security_profile_behaviors}
\title{Validates a Device Defender security profile behaviors specification}
\usage{
validate_security_profile_behaviors(
  behaviors = list(
    list(
      name = "string",
      metric = "string",
      criteria = list(
        comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set",
        value = list(
          count = 123,
          cidrs = list(
            "string"
          ),
          ports = list(
            123
          )
        ),
        durationSeconds = 123
      )
    )
  )
)
}
\arguments{
\item{behaviors}{[required] Specifies the behaviors that, when violated by a device (thing), cause an alert.}
}
\description{
Validates a Device Defender security profile behaviors specification.
}
\examples{

}
