% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{attach_thing_principal}
\alias{attach_thing_principal}
\title{Attaches the specified principal to the specified thing}
\usage{
attach_thing_principal(
  thingName = "string",
  principal = "string"
)
}
\arguments{
\item{thingName}{[required] The name of the thing.}

\item{principal}{[required] The principal, such as a certificate or other credential.}
}
\description{
Attaches the specified principal to the specified thing.
}
\examples{

}
