% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_ota_update}
\alias{create_ota_update}
\title{Creates an AWS IoT OTAUpdate on a target group of things or groups}
\usage{
create_ota_update(otaUpdateId, targets, files, roleArn, description = NULL,
  targetSelection = NULL, awsJobExecutionsRolloutConfig = NULL,
  additionalParameters = NULL)
}
\arguments{
\item{otaUpdateId}{The ID of the OTA update to be created.}

\item{targets}{The targeted devices to receive OTA updates.}

\item{files}{The files to be streamed by the OTA update.}

\item{roleArn}{The IAM role that allows access to the AWS IoT Jobs service.}

\item{description}{The description of the OTA update.}

\item{targetSelection}{Specifies whether the update will continue to run (CONTINUOUS), or will be complete after all the things specified as targets have completed the update (SNAPSHOT). If continuous, the update may also be run on a thing when a change is detected in a target. For example, an update will run on a thing when the thing is added to a target group, even after the update was completed by all things originally in the group. Valid values: CONTINUOUS | SNAPSHOT.}

\item{awsJobExecutionsRolloutConfig}{Configuration for the rollout of OTA updates.}

\item{additionalParameters}{A list of additional OTA update parameters which are name-value pairs.}
}
\description{
Creates an AWS IoT OTAUpdate on a target group of things or groups.
}
\examples{

}
