% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_security_profile}
\alias{create_security_profile}
\title{Creates a Device Defender security profile}
\usage{
create_security_profile(securityProfileName, behaviors,
  securityProfileDescription = NULL, alertTargets = NULL)
}
\arguments{
\item{securityProfileName}{The name you are giving to the security profile.}

\item{behaviors}{Specifies the behaviors that, when violated by a device (thing), cause an alert.}

\item{securityProfileDescription}{A description of the security profile.}

\item{alertTargets}{Specifies the destinations to which alerts are sent. (Alerts are always sent to the console.) Alerts are generated when a device (thing) violates a behavior.}
}
\description{
Creates a Device Defender security profile.
}
\examples{

}
