% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_thing_group}
\alias{create_thing_group}
\title{Create a thing group}
\usage{
create_thing_group(thingGroupName, parentGroupName = NULL,
  thingGroupProperties = NULL)
}
\arguments{
\item{thingGroupName}{[required] The thing group name to create.}

\item{parentGroupName}{The name of the parent thing group.}

\item{thingGroupProperties}{The thing group properties.}
}
\description{
Create a thing group.
}
\details{
This is a control plane operation. See \href{http://docs.aws.amazon.com/iot/latest/developerguide/authorization.html}{Authorization} for information about authorizing control plane actions.
}
\section{Accepted Parameters}{
\preformatted{create_thing_group(
  thingGroupName = "string",
  parentGroupName = "string",
  thingGroupProperties = list(
    thingGroupDescription = "string",
    attributePayload = list(
      attributes = list(
        "string"
      ),
      merge = TRUE|FALSE
    )
  )
)
}
}

