% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{create_topic_rule}
\alias{create_topic_rule}
\title{Creates a rule}
\usage{
create_topic_rule(
  ruleName = "string",
  topicRulePayload = list(
    sql = "string",
    description = "string",
    actions = list(
      list(
        dynamoDB = list(
          tableName = "string",
          roleArn = "string",
          operation = "string",
          hashKeyField = "string",
          hashKeyValue = "string",
          hashKeyType = "STRING"|"NUMBER",
          rangeKeyField = "string",
          rangeKeyValue = "string",
          rangeKeyType = "STRING"|"NUMBER",
          payloadField = "string"
        ),
        dynamoDBv2 = list(
          roleArn = "string",
          putItem = list(
            tableName = "string"
          )
        ),
        lambda = list(
          functionArn = "string"
        ),
        sns = list(
          targetArn = "string",
          roleArn = "string",
          messageFormat = "RAW"|"JSON"
        ),
        sqs = list(
          roleArn = "string",
          queueUrl = "string",
          useBase64 = TRUE|FALSE
        ),
        kinesis = list(
          roleArn = "string",
          streamName = "string",
          partitionKey = "string"
        ),
        republish = list(
          roleArn = "string",
          topic = "string"
        ),
        s3 = list(
          roleArn = "string",
          bucketName = "string",
          key = "string",
          cannedAcl = "private"|"public-read"|"public-read-write"|"aws-exec-read"|"authenticated-read"|"bucket-owner-read"|"bucket-owner-full-control"|"log-delivery-write"
        ),
        firehose = list(
          roleArn = "string",
          deliveryStreamName = "string",
          separator = "string"
        ),
        cloudwatchMetric = list(
          roleArn = "string",
          metricNamespace = "string",
          metricName = "string",
          metricValue = "string",
          metricUnit = "string",
          metricTimestamp = "string"
        ),
        cloudwatchAlarm = list(
          roleArn = "string",
          alarmName = "string",
          stateReason = "string",
          stateValue = "string"
        ),
        elasticsearch = list(
          roleArn = "string",
          endpoint = "string",
          index = "string",
          type = "string",
          id = "string"
        ),
        salesforce = list(
          token = "string",
          url = "string"
        ),
        iotAnalytics = list(
          channelArn = "string",
          channelName = "string",
          roleArn = "string"
        ),
        iotEvents = list(
          inputName = "string",
          messageId = "string",
          roleArn = "string"
        ),
        stepFunctions = list(
          executionNamePrefix = "string",
          stateMachineName = "string",
          roleArn = "string"
        )
      )
    ),
    ruleDisabled = TRUE|FALSE,
    awsIotSqlVersion = "string",
    errorAction = list(
      dynamoDB = list(
        tableName = "string",
        roleArn = "string",
        operation = "string",
        hashKeyField = "string",
        hashKeyValue = "string",
        hashKeyType = "STRING"|"NUMBER",
        rangeKeyField = "string",
        rangeKeyValue = "string",
        rangeKeyType = "STRING"|"NUMBER",
        payloadField = "string"
      ),
      dynamoDBv2 = list(
        roleArn = "string",
        putItem = list(
          tableName = "string"
        )
      ),
      lambda = list(
        functionArn = "string"
      ),
      sns = list(
        targetArn = "string",
        roleArn = "string",
        messageFormat = "RAW"|"JSON"
      ),
      sqs = list(
        roleArn = "string",
        queueUrl = "string",
        useBase64 = TRUE|FALSE
      ),
      kinesis = list(
        roleArn = "string",
        streamName = "string",
        partitionKey = "string"
      ),
      republish = list(
        roleArn = "string",
        topic = "string"
      ),
      s3 = list(
        roleArn = "string",
        bucketName = "string",
        key = "string",
        cannedAcl = "private"|"public-read"|"public-read-write"|"aws-exec-read"|"authenticated-read"|"bucket-owner-read"|"bucket-owner-full-control"|"log-delivery-write"
      ),
      firehose = list(
        roleArn = "string",
        deliveryStreamName = "string",
        separator = "string"
      ),
      cloudwatchMetric = list(
        roleArn = "string",
        metricNamespace = "string",
        metricName = "string",
        metricValue = "string",
        metricUnit = "string",
        metricTimestamp = "string"
      ),
      cloudwatchAlarm = list(
        roleArn = "string",
        alarmName = "string",
        stateReason = "string",
        stateValue = "string"
      ),
      elasticsearch = list(
        roleArn = "string",
        endpoint = "string",
        index = "string",
        type = "string",
        id = "string"
      ),
      salesforce = list(
        token = "string",
        url = "string"
      ),
      iotAnalytics = list(
        channelArn = "string",
        channelName = "string",
        roleArn = "string"
      ),
      iotEvents = list(
        inputName = "string",
        messageId = "string",
        roleArn = "string"
      ),
      stepFunctions = list(
        executionNamePrefix = "string",
        stateMachineName = "string",
        roleArn = "string"
      )
    )
  )
)
}
\arguments{
\item{ruleName}{[required] The name of the rule.}

\item{topicRulePayload}{[required] The rule payload.}
}
\description{
Creates a rule. Creating rules is an administrator-level action. Any user who has permission to create rules will be able to access data processed by the rule.
}
\examples{

}
