% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_authorizer}
\alias{delete_authorizer}
\title{Deletes an authorizer}
\usage{
delete_authorizer(authorizerName)
}
\arguments{
\item{authorizerName}{[required] The name of the authorizer to delete.}
}
\description{
Deletes an authorizer.
}
\section{Accepted Parameters}{
\preformatted{delete_authorizer(
  authorizerName = "string"
)
}
}

