% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{delete_stream}
\alias{delete_stream}
\title{Deletes a stream}
\usage{
delete_stream(
  streamId = "string"
)
}
\arguments{
\item{streamId}{[required] The stream ID.}
}
\description{
Deletes a stream.
}
\examples{

}
