% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_authorizer}
\alias{describe_authorizer}
\title{Describes an authorizer}
\usage{
describe_authorizer(authorizerName)
}
\arguments{
\item{authorizerName}{[required] The name of the authorizer to describe.}
}
\description{
Describes an authorizer.
}
\section{Accepted Parameters}{
\preformatted{describe_authorizer(
  authorizerName = "string"
)
}
}

