% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_billing_group}
\alias{describe_billing_group}
\title{Returns information about a billing group}
\usage{
describe_billing_group(
  billingGroupName = "string"
)
}
\arguments{
\item{billingGroupName}{[required] The name of the billing group.}
}
\description{
Returns information about a billing group.
}
\examples{

}
