% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_index}
\alias{describe_index}
\title{Describes a search index}
\usage{
describe_index(
  indexName = "string"
)
}
\arguments{
\item{indexName}{[required] The index name.}
}
\description{
Describes a search index.
}
\examples{

}
