% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_job}
\alias{describe_job}
\title{Describes a job}
\usage{
describe_job(jobId)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}
}
\description{
Describes a job.
}
\section{Accepted Parameters}{
\preformatted{describe_job(
  jobId = "string"
)
}
}

