% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{describe_thing_type}
\alias{describe_thing_type}
\title{Gets information about the specified thing type}
\usage{
describe_thing_type(thingTypeName)
}
\arguments{
\item{thingTypeName}{[required] The name of the thing type.}
}
\description{
Gets information about the specified thing type.
}
\section{Accepted Parameters}{
\preformatted{describe_thing_type(
  thingTypeName = "string"
)
}
}

