% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{detach_policy}
\alias{detach_policy}
\title{Detaches a policy from the specified target}
\usage{
detach_policy(policyName, target)
}
\arguments{
\item{policyName}{The policy to detach.}

\item{target}{The target from which the policy will be detached.}
}
\description{
Detaches a policy from the specified target.
}
\examples{

}
