% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{detach_thing_principal}
\alias{detach_thing_principal}
\title{Detaches the specified principal from the specified thing}
\usage{
detach_thing_principal(thingName, principal)
}
\arguments{
\item{thingName}{[required] The name of the thing.}

\item{principal}{[required] If the principal is a certificate, this value must be ARN of the certificate. If the principal is an Amazon Cognito identity, this value must be the ID of the Amazon Cognito identity.}
}
\description{
Detaches the specified principal from the specified thing.
}
\details{
This call is asynchronous. It might take several seconds for the detachment to propagate.
}
\section{Accepted Parameters}{
\preformatted{detach_thing_principal(
  thingName = "string",
  principal = "string"
)
}
}

