% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{get_ota_update}
\alias{get_ota_update}
\title{Gets an OTA update}
\usage{
get_ota_update(otaUpdateId)
}
\arguments{
\item{otaUpdateId}{The OTA update ID.}
}
\description{
Gets an OTA update.
}
\examples{

}
