% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_ota_updates}
\alias{list_ota_updates}
\title{Lists OTA updates}
\usage{
list_ota_updates(
  maxResults = 123,
  nextToken = "string",
  otaUpdateStatus = "CREATE_PENDING"|"CREATE_IN_PROGRESS"|"CREATE_COMPLETE"|"CREATE_FAILED"
)
}
\arguments{
\item{maxResults}{The maximum number of results to return at one time.}

\item{nextToken}{A token used to retrieve the next set of results.}

\item{otaUpdateStatus}{The OTA update job status.}
}
\description{
Lists OTA updates.
}
\examples{

}
