% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_principal_things}
\alias{list_principal_things}
\title{Lists the things associated with the specified principal}
\usage{
list_principal_things(nextToken = NULL, maxResults = NULL, principal)
}
\arguments{
\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return in this operation.}

\item{principal}{[required] The principal.}
}
\description{
Lists the things associated with the specified principal.
}
\section{Accepted Parameters}{
\preformatted{list_principal_things(
  nextToken = "string",
  maxResults = 123,
  principal = "string"
)
}
}

