% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{list_things_in_thing_group}
\alias{list_things_in_thing_group}
\title{Lists the things in the specified group}
\usage{
list_things_in_thing_group(thingGroupName, recursive = NULL,
  nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{thingGroupName}{The thing group name.}

\item{recursive}{When true, list things in this thing group and in all child groups as well.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the things in the specified group.
}
\examples{

}
