% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{search_index}
\alias{search_index}
\title{The query search index}
\usage{
search_index(indexName = NULL, queryString, nextToken = NULL,
  maxResults = NULL, queryVersion = NULL)
}
\arguments{
\item{indexName}{The search index name.}

\item{queryString}{[required] The search query string.}

\item{nextToken}{The token used to get the next set of results, or \strong{null} if there are no additional results.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{queryVersion}{The query version.}
}
\description{
The query search index.
}
\section{Accepted Parameters}{
\preformatted{search_index(
  indexName = "string",
  queryString = "string",
  nextToken = "string",
  maxResults = 123,
  queryVersion = "string"
)
}
}

