% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{set_logging_options}
\alias{set_logging_options}
\title{Sets the logging options}
\usage{
set_logging_options(loggingOptionsPayload)
}
\arguments{
\item{loggingOptionsPayload}{The logging options payload.}
}
\description{
Sets the logging options.
}
\details{
NOTE: use of this command is not recommended. Use \code{SetV2LoggingOptions} instead.
}
\examples{

}
