% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{set_v2_logging_level}
\alias{set_v2_logging_level}
\title{Sets the logging level}
\usage{
set_v2_logging_level(
  logTarget = list(
    targetType = "DEFAULT"|"THING_GROUP",
    targetName = "string"
  ),
  logLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED"
)
}
\arguments{
\item{logTarget}{[required] The log target.}

\item{logLevel}{[required] The log level.}
}
\description{
Sets the logging level.
}
\examples{

}
