% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_event_configurations}
\alias{update_event_configurations}
\title{Updates the event configurations}
\usage{
update_event_configurations(
  eventConfigurations = list(
    list(
      Enabled = TRUE|FALSE
    )
  )
)
}
\arguments{
\item{eventConfigurations}{The new event configuration values.}
}
\description{
Updates the event configurations.
}
\examples{

}
