% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_role_alias}
\alias{update_role_alias}
\title{Updates a role alias}
\usage{
update_role_alias(roleAlias, roleArn = NULL,
  credentialDurationSeconds = NULL)
}
\arguments{
\item{roleAlias}{The role alias to update.}

\item{roleArn}{The role ARN.}

\item{credentialDurationSeconds}{The number of seconds the credential will be valid.}
}
\description{
Updates a role alias.
}
\examples{

}
