% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{update_thing_groups_for_thing}
\alias{update_thing_groups_for_thing}
\title{Updates the groups to which the thing belongs}
\usage{
update_thing_groups_for_thing(thingName = NULL,
  thingGroupsToAdd = NULL, thingGroupsToRemove = NULL)
}
\arguments{
\item{thingName}{The thing whose group memberships will be updated.}

\item{thingGroupsToAdd}{The groups to which the thing will be added.}

\item{thingGroupsToRemove}{The groups from which the thing will be removed.}
}
\description{
Updates the groups to which the thing belongs.
}
\section{Accepted Parameters}{
\preformatted{update_thing_groups_for_thing(
  thingName = "string",
  thingGroupsToAdd = list(
    "string"
  ),
  thingGroupsToRemove = list(
    "string"
  )
)
}
}

