% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot_operations.R
\name{validate_security_profile_behaviors}
\alias{validate_security_profile_behaviors}
\title{Validates a Device Defender security profile behaviors specification}
\usage{
validate_security_profile_behaviors(behaviors)
}
\arguments{
\item{behaviors}{Specifies the behaviors that, when violated by a device (thing), cause an alert.}
}
\description{
Validates a Device Defender security profile behaviors specification.
}
\examples{

}
